+++ CHANGELOG.md
@@ -6,7 +6,13 @@ and this project attempts to adhere to Semantic Versioning.
 
 ## UNRELEASED
 
+### Changed
 
+* no longer delete signoffs (keep all and return most recent)
+
+### Added
+
+* a changelog for signoff modifications
 
 ## 3.14.1: [2026-01-30]
 
@@ -16,13 +22,18 @@ and this project attempts to adhere to Semantic Versioning.
 
 ## 3.14.0: [2026-01-26]
 
+
 ### Fixed
 
 * added documentation for the case archive resume-archiving endpoint
 
+
 ### Changed
 
 * added "project" archive support in addition to case archives
+
+### Added
+* changelog for signoffs (no front end changes)
 
 ## 3.13.1: [2025-11-13]